Revert "Refactor `serialize` function"
This reverts commit 184f0cf428d317da7aeb9d6f92a0aa92776401b6.
