+++ crates/sncast/src/helpers/mod.rs
@@ -1,5 +1,4 @@
 pub mod account;
-pub mod arguments;
 pub mod block_explorer;
 pub mod braavos;
 pub mod config;
