+++ crates/sncast/src/main.rs
@@ -3,11 +3,10 @@ use crate::starknet_commands::{
     account, account::Account, call::Call, declare::Declare, deploy::Deploy, invoke::Invoke,
     multicall::Multicall, script::Script, show_config::ShowConfig, tx_status::TxStatus,
 };
-use anyhow::{Context, Result};
-use data_transformer::reverse_transform_output;
+use anyhow::{Context, Result, bail};
+use data_transformer::{reverse_transform_output, transform};
 use foundry_ui::{Message, UI};
 use sncast::helpers::account::generate_account_name;
-use sncast::helpers::arguments::Arguments;
 use sncast::helpers::output_format::output_format_from_json_flag;
 use sncast::response::call::CallResponse;
 use sncast::response::cast_message::SncastMessage;
@@ -16,9 +15,10 @@ use sncast::response::declare::DeclareResponse;
 use sncast::response::errors::ResponseError;
 use sncast::response::explorer_link::{ExplorerLinksMessage, block_explorer_link_if_allowed};
 use sncast::response::transformed_call::TransformedCallResponse;
-use std::io;
 use std::io::IsTerminal;
+use std::{fs, io};
 
+use crate::starknet_commands::deploy::DeployArguments;
 use camino::Utf8PathBuf;
 use clap::{CommandFactory, Parser, Subcommand};
 use configuration::load_config;
@@ -155,6 +155,74 @@ enum Commands {
 
     /// Serialize Cairo expressions into calldata
     Serialize(Serialize),
+}
+
+#[derive(Debug, Clone, clap::Args)]
+#[group(multiple = false)]
+pub struct Arguments {
+    /// Arguments of the called function serialized as a series of felts
+    #[arg(short, long, value_delimiter = ' ', num_args = 1..)]
+    pub calldata: Option<Vec<String>>,
+
+    // Arguments of the called function as a comma-separated string of Cairo expressions
+    #[arg(long, allow_hyphen_values = true)]
+    pub arguments: Option<String>,
+}
+
+impl Arguments {
+    fn try_into_calldata(
+        self,
+        contract_class: Option<ContractClass>,
+        selector: &Felt,
+        abi_file: Option<Utf8PathBuf>,
+    ) -> Result<Vec<Felt>> {
+        if let Some(calldata) = self.calldata {
+            return calldata
+                .iter()
+                .map(|data| {
+                    Felt::from_dec_str(data)
+                        .or_else(|_| Felt::from_hex(data))
+                        .context("Failed to parse to felt")
+                })
+                .collect();
+        }
+
+        let abi: Vec<AbiEntry> = match (contract_class, abi_file) {
+            (Some(_), Some(_)) => {
+                bail!("`contract_class` and `abi_file` params are mutually exclusive");
+            }
+            (Some(ContractClass::Sierra(sierra_class)), None) => {
+                serde_json::from_str(sierra_class.abi.as_str())
+                    .context("Couldn't deserialize ABI received from network")?
+            }
+            (Some(_), None) => {
+                bail!("Transformation of arguments is not available for Cairo Zero contracts");
+            }
+            (None, Some(path)) => {
+                let abi_str = fs::read_to_string(path).context("Failed to read ABI file")?;
+                serde_json::from_str(&abi_str).context("Failed to deserialize ABI from file")?
+            }
+            (None, None) => {
+                bail!("Either `contract_class` or `abi_file` must be provided");
+            }
+        };
+
+        let args = self.arguments.unwrap_or_default();
+        transform(&args, &abi, selector).context("Failed to transform arguments into calldata")
+    }
+}
+
+impl From<DeployArguments> for Arguments {
+    fn from(value: DeployArguments) -> Self {
+        let DeployArguments {
+            constructor_calldata,
+            arguments,
+        } = value;
+        Self {
+            calldata: constructor_calldata,
+            arguments,
+        }
+    }
 }
 
 fn main() -> Result<()> {
@@ -383,8 +451,34 @@ async fn run_async_command(cli: Cli, config: CastConfig, ui: &UI) -> Result<()> 
         }
 
         Commands::Serialize(serialize) => {
-            let result = starknet_commands::serialize::serialize(serialize, config, ui)
-                .await
+            let Serialize {
+                function,
+                arguments,
+                rpc,
+                abi_file,
+                ..
+            } = serialize.clone();
+
+            let provider = rpc.get_provider(&config, ui).await?;
+
+            let selector = get_selector_from_name(&function)
+                .context("Failed to convert entry point selector to FieldElement")?;
+
+            let arguments = Arguments {
+                calldata: None,
+                arguments: Some(arguments),
+            };
+            let calldata = if let Some(abi_file) = abi_file {
+                arguments
+                    .try_into_calldata(None, &selector, Some(abi_file))
+                    .context("Failed to transform arguments into calldata")?
+            } else {
+                let class_hash = serialize.class_hash(config, ui).await?;
+                let contract_class = get_contract_class(class_hash, &provider).await?;
+                arguments.try_into_calldata(Some(contract_class), &selector, None)?
+            };
+
+            let result = starknet_commands::serialize::serialize(calldata)
                 .map_err(handle_starknet_command_error);
 
             process_command_result("serialize", result, ui, None);
